+++ bundle/src/main/java/com/adobe/acs/commons/httpcache/engine/impl/delegate/HttpCacheEngineBindingsDelegate.java
@@ -174,7 +174,7 @@ public class HttpCacheEngineBindingsDelegate {
     }
 
     public List<HttpCacheConfig> getCacheConfigs() {
-        return cacheConfigs;
+        return Collections.unmodifiableList(cacheConfigs);
     }
 
     public Map<String, HttpCacheStore> getCacheStoresMap() {